# Improve performance, upgrade to Netty `4`