--- a/project.clj
+++ b/project.clj
@@ -5,7 +5,7 @@
             :url "https://www.gnu.org/licenses/old-licenses/gpl-2.0.txt"}
   :dependencies [[org.clojure/clojure "1.11.1"]
                  [org.clojure/core.async "1.6.673"]
-                 [org.jboss.netty/netty "3.2.7.Final"]
+                 [io.netty/netty-all "4.1.97.Final"]
                  [org.clojure/data.finger-tree "0.0.1"]
                  [com.clojure-goes-fast/clj-async-profiler "1.0.5"]]
   :main ^:skip-aot data-store.core