--- a/src/data_store/core.clj
+++ b/src/data_store/core.clj
@@ -3,12 +3,7 @@
   (:require
    [data-store.handler :refer [handler]]
    [data-store.server :refer [serve!]]
-   [data-store.store :refer [dynamic-expiry!]]
-   [clj-async-profiler.core :refer [profile]]))
-
-(defn profile-main []
-  (profile
-   (serve! 6379 handler)))
+   [data-store.store :refer [dynamic-expiry!]]))
 
 (defn -main []
   (serve! 6379 handler)