--- a/src/data_store/core.clj
+++ b/src/data_store/core.clj
@@ -1,11 +1,11 @@
 (ns data-store.core
   (:gen-class)
   (:require
-   [data-store.handler :refer [output-handler-strategy
-                               store-handler-strategy]]
+   [data-store.handler :refer [output-handler-factory
+                               store-handler-factory]]
    [data-store.server :refer [serve!]]
    [data-store.store :refer [dynamic-expiry!]]))
 
 (defn -main []
-  (serve! 6379 output-handler-strategy store-handler-strategy)
+  (serve! 6379 output-handler-factory store-handler-factory)
   (dynamic-expiry! 100))