--- a/src/data_store/frame.clj
+++ b/src/data_store/frame.clj
@@ -53,6 +53,4 @@
     (vector? x)
     (let [length (count x)]
       (str "*" length delim (join (map serialize x))))
-    (contains? x :error)
-    (str "-" (:error x) delim)
     :else {:error "Could not serialize"}))