--- a/src/data_store/handler.clj
+++ b/src/data_store/handler.clj
@@ -4,68 +4,42 @@
 
 (def delim "\r\n")
 
-(defn command-store-get [input store time]
+(defn command-get [input store time]
   (let [store-key (first input)]
     (cond
       (not store-key)
-      store
+      [store (str "-Error nothing to get" delim)]
       (not (contains? store store-key))
-      store
+      [store (serialize nil)]
       (not (string? (:val (get store store-key))))
-      store
+      [store (str "-Error not a string" delim)]
       :else (let [exp (:exp (get store store-key))
                   expired? (and exp (>= time exp))]
               (if
                expired?
-                (dissoc store store-key)
-                store)))))
+                [(dissoc store store-key) (serialize nil)]
+                [store (serialize (:val (get store store-key)))])))))
 
-(defn command-output-get [input store time]
-  (let [store-key (first input)]
-    (cond
-      (not store-key)
-      (str "-Error nothing to get" delim)
-      (not (contains? store store-key))
-      (serialize nil)
-      (not (string? (:val (get store store-key))))
-      (str "-Error not a string" delim)
-      :else (let [exp (:exp (get store store-key))
-                  expired? (and exp (>= time exp))]
-              (if
-               expired?
-                (serialize nil)
-                (serialize (:val (get store store-key))))))))
-
-(defn command-store-set [input store time]
+(defn command-set [input store time]
   (let [k (first input)
         v (second input)]
     (cond
       (or (not k) (not v))
-      store
+      [store (str "-Error nothing to set" delim)]
       (not (string? v))
-      store
+      [store (str "-Error not a string" delim)]
       :else (let [sub-command (nth input 2 nil)
                   sub-command-arg (nth input 3 nil)]
               (cond
                 (and (= sub-command "EX") sub-command-arg)
-                (into store {k {:val v :exp (+ time (* 1000 (Integer/parseInt sub-command-arg)))}})
+                [(into store {k {:val v :exp (+ time (* 1000 (Integer/parseInt sub-command-arg)))}}) (serialize "OK")]
                 (and (= sub-command "PEX") sub-command-arg)
-                (into store {k {:val v :exp (+ time (Integer/parseInt sub-command-arg))}})
+                [(into store {k {:val v :exp (+ time (Integer/parseInt sub-command-arg))}}) (serialize "OK")]
                 (and (= sub-command "EXAT") sub-command-arg)
-                (into store {k {:val v :exp (* 1000 (Integer/parseInt sub-command-arg))}})
+                [(into store {k {:val v :exp (* 1000 (Integer/parseInt sub-command-arg))}}) (serialize "OK")]
                 (and (= sub-command "PXAT") sub-command-arg)
-                (into store {k {:val v :exp (Integer/parseInt sub-command-arg)}})
-                :else (into store {k {:val v}}))))))
-
-(defn command-output-set [input _ _]
-  (let [k (first input)
-        v (second input)]
-    (cond
-      (or (not k) (not v))
-      (str "-Error nothing to set" delim)
-      (not (string? v))
-      (str "-Error not a string" delim)
-      :else (serialize "OK"))))
+                [(into store {k {:val v :exp (Integer/parseInt sub-command-arg)}}) (serialize "OK")]
+                :else [(into store {k {:val v}}) (serialize "OK")])))))
 
 (defn command-expire [input store time]
   (let [store-key (first input)
@@ -129,24 +103,13 @@
       [(into store {key (into (get store key {}) {:val new-val})}) (serialize (count new-val))])
     [store (str "-Error nothing to push" delim)]))
 
-(defn command-output-unknown []
-  (serialize "OK"))
+(defn command-unknown [_ store _]
+  [store (serialize "OK")])
 
-(defn command-store-unknown [_ store _]
-  store)
-
-(defn store-handler [command & args]
-  (if-let [dispatch-handler (ns-resolve 'data-store.handler (symbol (str "command-store-" command)))]
+(defn command-handler [command & args]
+  (if-let [dispatch-handler (ns-resolve 'data-store.handler (symbol (str "command-" command)))]
     (apply dispatch-handler args)
-    (apply command-store-unknown args)))
+    (apply command-unknown args)))
 
-(defn store-handler-adapter [store input time]
-  (store-handler (lower-case (first input)) (rest input) store time))
-
-(defn output-handler [command & args]
-  (if-let [dispatch-handler (ns-resolve 'data-store.handler (symbol (str "command-output-" command)))]
-    (apply dispatch-handler args)
-    (command-output-unknown)))
-
-(defn output-handler-adapter [store input time]
-  (output-handler (lower-case (first input)) (rest input) store time))
+(defn handler [store input time]
+  (command-handler (lower-case (first input)) (rest input) store time))