* Upgrade to [Netty 4](https://netty.io/wiki/user-guide-for-4.x.html)
* Improve performance, like using multimethods instead of function lookups, and `(nth n 0)` instead of `(first n)`

## Before

```sh
$ redis-benchmark -t set,get, -n 1000000 -q

SET: 48262.55 requests per second, p50=0.447 msec                   
GET: 86021.51 requests per second, p50=0.303 msec 
```

## After

```sh
$ redis-benchmark -t set,get, -n 1000000 -q

SET: 79302.14 requests per second, p50=0.303 msec                   
GET: 83430.67 requests per second, p50=0.311 msec    
```